RestApiArticlesApp/requirements.txt
Django==4.1
djangorestframework==3.13.1
flake8==5.0.4
